Fix SQL injection at function source tiles url
Since the end user controls the query json that is
passed to postgres, it is possible to perform
a SQL injection on the function source tiles endpoint.
By preparing the query we avoid such issues. However
we still need to sanitize by hand the schema and function
parameters since they can't be a part of the prepared query.

Using prepared queries also forces us to use the right
data types when passing parameters to postgres. As such,
the coordinates system that were using unsigned
integers uses signed integers since they map directly
to postgres types.

Thanks to Frédéric Rodrigo for discovering this issue.