[WIP] Upgrade actix dependencies
Bump actix dependencies to the latest versions

The server.rs `fn new(pool, config)` needs more work because the new `HttpServer::new` returns a future that needs to be awaited.  I am not an expert in actix, and I'm not certain why `actix::System::new()` was being returned before, and if its still needed.  Some help is welcome.