Feat: max_feature_count query limit
Implements https://github.com/maplibre/martin/issues/384 - ability to limit the number of features included in a tile from a Postgres table/view.

This allows viewing zoomed-out view of a table with a reasonable speed.

If set on a CLI, overrides whatever is set in the config file (if given).