Partial implementation of the id_column detection
Detecting id column should only be done when adding all auto-detected tables to the sources.

I rewrote the logic, but did not test it fully. I added a simple test (we may want to add a few more), but currently tests break - not certain why yet - maybe the `from_schema` is not working correctly?

It's getting a bit late - try going from here.