[WIP] Fixing comment bug
When comment is set on a table or a function, and that tbl/func is pre-configured as part of the config file, the SQL comment is currently ignored.

This commit does NOT fix the issue, only identifies it and adds an integration test. Note that it will pass at the moment, and once fixed, the expected result files will change.