Add basic cog support (#1590)
Try to fix #875 . Still a draft and a lot to discuss and finish.


- [ ] Remote file on like S3. We will make it in next PRs. Let's make
this PR just a basic start.
- [x] Made the padded part transparent
- [x] Made the [gdal
nodata](https://gdal.org/en/stable/drivers/raster/gtiff.html#nodata-value)
transparent
- [x] Add tests
- [x] Add doc
- [x] Support rgb/rgba u8 tif tiled image. We will do others in next PRs
and leave this PR simple.

## Color types and bits per smaple
|colory type|bits per sample|supported|status|
|----|----|----|----|
|rgb/rgba|8|✅||
|rgb/rgba|16/32...|🛠️|working on, will be added in next PRs|
|gray|8/16/32...|🛠️|working on, will be added in next PRs|

## Comporession
The crate we used support these methods.
| Compression Format | tiff crate(This PR based on this) | GDAL creation
optinos |
|--------------------|------------------|----------------|
| None               | ✓                | ✓              |
| LZW                | ✓                | ✓ (gdal default)             |
| JPEG               |                  | ✓              |
| Deflate            | ✓                | ✓              |
| ZSTD               |                  | ✓              |
| WEBP               |                  | ✓              |
| LERC               |                  | ✓              |
| LERC_DEFLATE       |                  | ✓              |
| LERC_ZSTD          |                  | ✓              |
| LZMA               |                  | ✓              |
| PackBits | ✓ | Not sure but couldn't find this on gdal doc |

## Some terms and ref links
[7.1.2 Tiles in COG
spec](https://docs.ogc.org/is/21-026/21-026.html#_tiles)
[Tiff 6.0 spec page
67](https%3A%2F%2Fwww.itu.int%2Fitudoc%2Fitu-t%2Fcom16%2Ftiff-fx%2Fdocs%2Ftiff6.pdf)
[Chapter 4 of COG
spec](https://docs.ogc.org/is/21-026/21-026.html#_terms_and_definitions)

## Tile and padding
> In the context for a TIFF file, Tiling is a strategy for dividing the
content in the TIFF file differently than using the classical Strips.
Tiles, as defined in the TIFF version 6.0 specification, can be mapped
to the ones defined in the OGC Two Dimensional Tile Matrix Set Standard
(2D-TMS). For example in 2D-TMS, the TIFF 6.0 forces all tiles to be of
the same size. This is possible with the introduction of the concept of
padding: if necessary extra blank rows or columns are added to the
right-most and bottom-most tile to make them the same shape as other
tiles. However, the naming of the TIFF tags used version 6.0 and the
property names used in the 2D-TMS differ. The following table provides a
mapping between the two standards.

| OGC 2D-TMS | TIFF v. 6.0 | Definition |

|---------------|--------------|-------------------------------------------------|
| TileWidth | TileWidth | The tile width in pixels. The number of
columns in each tile |
| TileHeight | TileLength | The tile height in pixels. The number of
rows in each tile |
| MatrixWidth | TilesAcross | Number of tiles in the width direction |
| MatrixHeight | TilesDown | Number of tiles in the height direction

---------

Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>