fix: `trivially_copy_pass_by_ref` lint (#1654)
As noticed in https://github.com/maplibre/martin/pull/1649, this lint
fails if chainging the default and thus making it run.

For `Copy`-enums this should basically be the same thing, should as far
as I understood not even be an API-breaking change.

Perf-wise, the compiler should be able to optimise it the same way in
the end
=> sadly likely won't have a performance impact