fix: abandoned changelog.md being accidentally chosen as a source for release-notes (#1665)
both `body_part` and `draft` override the current text.
That is really annoying and must be a bug, but I don't think we need
this functionality (do we?)

I don't know if we want `generate_release_notes`. I have added it, but
we might not want this..

(and I should have used my fork, sorry)