Fix martin-ui crate packaging
The martin-ui dir must be part of the martin package, otherwise it does not get included in the `cargo publish -p martin`.  We could also do it with a symlink from inside the `/martin` dir, but that gets a bit confusing - easier to just keep ui inside martin itself, as it is not used anywhere outside of that crate.