Fix martin-ui crate packaging (#1668)
The martin-ui dir must be part of the martin package, otherwise it does
not get included in the `cargo publish -p martin`. We could also do it
with a symlink from inside the `/martin` dir, but that gets a bit
confusing - easier to just keep ui inside martin itself, as it is not
used anywhere outside of that crate.

In order to prevent these bugs in the future, we might want to run
`cargo +nightly publish --workspace -Z package-workspace --dry-run` as
part of some CI -- this will take into account that some crates depend
on the other, and that some/all of them have not been published yet

fixes #1667

---------

Co-authored-by: Frank Elsinga <frank@elsinga.de>