2024 rust edition (#1698)
This PR upgrades (`cargo fix --eddition`, then fix what does not make
sense) our crates to the 2024 edition.

While quite early, this enables things such as the v3 resolver, which
might be helpfull to prevent MRSV breakages in the future by a MSRV
breakage now.

Most of the changes are from rustfmt => going commit-by-commit is likely
a good approach