feat: `mbtiles` hash extraction support (#1787)
(PR is based on #1786 => merge said PR first before looking at this
one.)

This PR adds support to the `mbtiles` crate to hand out which hash
corresponds to a tile.

The flat case might be debatable: Here I chose to md5 hash it to make
upstream code more readable. The alternative would be to in this case
return `None` for the hash.

This looks more intimidating than it actully is. Most of the lines are
unit-tests.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>