feat: add simpler etag handling to all non-critical endpoints (#1836)
This PR adds etag handling via an actix wrap instruction to all
"non-tile" endpoints.
I ommitted some endpoints which I suspect are not called more than once
such as `/` if the webui is not compiled in.

The reason for not being smarter for some is that honestly, the impact
of this is less than I expected based on benchmarks in #1834 and that
`actix-middleware-etag` uses the same hash algorithm.

<del>
I have not addded tests for this as frankly, this does not seem like a
feature that needs deep tests.
</del>
By request, this PR now includes all stemi-static headers.

---------

Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>
Co-authored-by: sharkAndshark <zhangyijunmetro@hotmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>