feat(ci): add attestation and sbom generation to the docker images (#1878)
> [!IMPORTANT]
> this PR needs https://github.com/maplibre/martin/pull/1881 to function
and thus has it merged in
> => review said PR first to reduce this PRs diff ^^


This PR adds attestation and sbom generation to the docker images.

This allows our users to verify their "authenticity" and better inspect
what is in them.
I doubt that a lot of people actually do this though ^^
=> https://github.com/maplibre/martin/attestations

But it is nice to have and should not cause any maintenance overhead. At
least it never caused any on the other repos where I enabled it ^^