chore(ci): reduce CI-Time by splitting cross to its own step of the CI build (#1902)
This PR moves cross to its own step in CI which others (lamda,
packaging, docker) can depend on.
It is a hard dependecy of https://github.com/maplibre/martin/pull/1897
as without refactoring this, changing cross for something running a
modern node version is quite hard.

As discussed in https://github.com/maplibre/martin/pull/1876, we
currently build docker files and two static binaries in the same step.

This is excessively slow (could be paralell) and we need the musl
binaries for other steps anyway:

| previously | now |
|--------|--------|
| [32m 20s](https://github.com/maplibre/martin/actions/runs/16037935228)
| [19m
26s](https://github.com/maplibre/martin/actions/runs/16073125582/) |

Moving this step outside is likely a good step.

> [!NOTE]
> We can shave off a few minutes by.. not using cross, but this is a
different PR.

This was the dependency tree beforehand:

```mermaid
flowchart TD
    docker -- "linux-musl ONLY amd" --> lamda
    docker -- "linux-musl" --> package
    build -- "linux-gnu" --> package
    build -- debian --> package & test-with-service
    build -- macos --> package
    build -- windows ONLY amd --> package & test-multi-os
    build -- "linux-gnu ONLY amd" --> test-with-service & test-multi-os
    test-multi-os --> package
    test-with-service --> package
```

The proposed dependency tree:

```mermaid
flowchart TD
    build-musl -- "linux-musl ONLY amd" --> lamda
    build-musl -- "linux-musl" --> package & docker
    build -- "linux-gnu" --> package
    build -- debian --> package & test-with-service
    build -- macos --> package
    build -- windows ONLY amd --> package & test-multi-os
    build -- "linux-gnu ONLY amd" --> test-with-service & test-multi-os
    test-multi-os --> package
    test-with-service --> package
    lamda --> package
```

---------

Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>