chore: move from eslint to biomejs for formatting/linting (#1909)
this breaks out another of the components (which are not blocked by node
compatability) of #1897

The PR is more of a preference, but one I would really like to see.
biomejs does have a bit of functionality (such as sorting json and
imports) which we currently do manually.

The linting and formatting is also much much faster than the laternative
JS-based linters.

The demo frontend does not quite clearing the type-check and also not
quite clearing all lints, but it also did not clear that beforehand. No
change there.
I fixed a few of the lower hanging fruits

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>