chore(ci): make sure that rustup update is called at the start of CI steps (#1934)
During FOSS4G I and @nyurik talked and he brought up that our rust
version not being well defined is a think which might fuck us over some
time in the future.

We should have a more consistent rust version in CI.

And yes, this is not strictly the solution.. but hey.. what is the
likelyhood of a rust release happening during the 20m of the CI =>
fairly low.