chore(ci): don't cache `cargo build` (#1935)
This is another artefact of last week.
We examined if we are cache-efficent and given that we are trying to
push a 50GB cube into a 10GB hole.. => hell no.

We need to reduce where we cache to a minimum, i.e. where it is in the
critical path and blocking.
Since we need to reduce, it makes sense to take this away from the
builds and keep it for the unit tests.