chore(core): make sure that dependencys are clearly attributed (#1947)
During the discovery-PR of what the next best step in pulling into
`martin-core` is, I discovered, that our dependency-feature gating could
be tighter.

The primary motivation behind this PR is that if what feature has which
dependency is well-defined, migration is much simpler.

This PR was tested via 

```bash
cargo hack check --each-feature --keep-going --locked
```