feat(ui): Add an analytics section via `/_/metrics` (#1942)
This draft breaks out just the analytics section of the frontend

It is a bit larger, but most of the changes are tests, so likey fine.
The big part is tests.

The commits on this one are a bit of a mess, but I am scared to squash
at this point. The rebase would likely result in at least one bug.

# Screenshot

<img width="1444" height="762" alt="image"
src="https://github.com/user-attachments/assets/e6fbd312-b00c-4ba1-927c-b482c958b7dc"
/>

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>