fix(ci): re-enable caches for `x86_64-apple-darwin` (#1970)
I missed this in the previous PR.

I just assumed that all runners are roughly equivalently fast.
They are not. Not by a long shot. MacOS x86 is lagging behind the others
by a lot.

In the previous PR I somehow did not noticed this.

If we enable Caching just for this one, our CI speeds up again.