fix(ui): improve subpath deployment support (#1978)
This improves https://github.com/maplibre/martin/issues/1977 a bit and
makes this at least work on the baseline for people running martin on a
base-path:

```diff
-      - "traefik.http.routers.navigatum-martin.rule=Host(`nav.tum.de`) && (PathPrefix(`/tiles`) || PathPrefix(`/_/`) || PathPrefix(`/catalog`))"
+      - "traefik.http.routers.navigatum-martin.rule=Host(`nav.tum.de`) && (PathPrefix(`/tiles`) || PathPrefix(`/_/`))"
```

The issue is mainly, that otherwise one would need to confiugre every
source in traefik.. not pheasable for lots of our users.
By using `href` instead of `origin` this works as users expect.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>