feat(martin-cp): add a warning if `--concurrency 1` and an error if `--concurrency 0` (#2027)
This PR implements a slightly more smart algorithm for detecting if we
should warn of `--concurrency 1`.

If we are copying locally, having more threads is likely not very
effective, unless we have to re-encode.
I am not warning of the re-encoding case to not overcomplicate things
for this PR review wise.

<img width="1466" height="646" alt="image"
src="https://github.com/user-attachments/assets/9be64100-ebf2-4aa5-bf8d-f9f27cbcfd47"
/>

Becausse of the nature of this functionality, I did not attach tests

Resolves #1990

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>