chore(deps): pin actions to immutable shas instead of versions (#2074)
This PR pins all of our action dependencys to shas instead of versions.

This is another measure to increase our security on the OpenSFF
Scorecard.

This migration was performed via the pin-github-action tool and I don't
see a huge argument against this as github keeps pushing back immutable
actions for a year now.
Even when they finally release said feature, this will not affect all
actions at it seems.