chore(releases): enable generating SLSA build provenance attestation (#2073)
Since we build our releases in CI, we might as well attest them.
This should give us a few points on the openssf's security scorecard, as
far as I can tell.

I am not sure if we have to download the attestation and attach it to
the release, but since this requires a release, we will have to check it
out in the future.