chore(core): migrate the configuration to be core capable (#2104)
Currently, a thing that couples our achitecture to the core part is the
configuration subsystem.
So currently, our architecture is that every distinct part has its
configuration, which lives there. How to get from the config to the
actual running code is not very consistent.

My plan is that:
- core has "things that could be usefull for others" => the stable core
=> `Sorce` implementations, core traits, ...
- while confiuration is tightly coupled to us => not in the particularly
usefull for others => how we configure from the config file concretely


The PR somewhat grew over this whole refactoring.
It is mostly to discuss the idea, I can break this apart further =>
staying reviewable.