chore(deps): try to use githubs version shas instead of shortened (#2144)
This possibly fixes Dependabot not updating the comments of GitHub shas
correctly.

Honestly, the best cause of action is to see if Dependabot misbehaves
again.
Thanks for notifying me.
Based on https://github.com/community/maintainers/discussions/603 using
full tags is the canonicaly-accepted answer.

Example of this misbehaving:
- https://github.com/maplibre/maplibre-gl-js/pull/6352/files

This might be coming from being too literal in the translation.

Further context:
- https://github.com/dependabot/dependabot-core/issues/7912