chore(core): change the errors of the Source trait to be dyn based (#2158)
The `Source` trait currently returns `MartinError`s.
This ties this trait to the `martin`-crate as I don't want to move the
errors.

Moving the erros would be pointless as others cannot add an error here.
I benchmarked the performance with the existing bench and the
performance is about the same (-1.5%, but that is likely just noise)

<img width="966" height="537" alt="image"
src="https://github.com/user-attachments/assets/74435584-b1df-4a9d-9c35-86e0fe233b2e"
/>