chore(core) final push in the pg-move (#2192)
This is the final push in the postgres move to core.

For the function/table resolver, this needs a bit of an refactoring
before these would be publicly usefull. I am clasifying them as
"configuration"-stuff and moving them to the config-file area.

I am not sure if this is the final resting place, but for now this seems
an OK place to push this to.

This PR also fixes two unrelated issues:
- some of our constants are `static` instead of `const`
- the `postgres` configuration module (in the config file "postgres") is
named `pg`