chore(security): add a colldown setting for dependency updates (#2206)
this PR adds a colldown setting for how we dependeny update.
This increases the likelyhood of our security tooling picking up on
vulnerabilities before they are in master.

Given that we are doing monthly updates, we run the risk of not getting
an update for another month, but that is not really avoidable unless we
update more frequently