Allow test.sh to run on darwin (#2200)
This allowed me to run integration tests locally. Mainly it is
portability around `sed` and `tr`, and allowing darwin to use `ogrmerge`
(will still require that setup just like linux) and `vtzero`.

---

There is another OS check in the `justfile` which ultimately diffs the
files after the script runs. I haven't updated that yet, as
most everything there looked fine, but I was seeing some json header
diffs. Namely, the expected is brotli encoding whereas the default is
gzip.

I didn't see anything in the test suite setup (such as args, or
`tests/config.yaml`) that would be setting `preferred_encoding` to
brotli, so I'm wondering why that is expected to be the case as opposed
to the default gzip. Was it never `bless`ed? Am I missing somewhere in
the test config where that is being changed?

---------

Co-authored-by: Frank Elsinga <frank@elsinga.de>