chore(tests): improve the style fixutres (#2229)
Currently two of the styles don't work locally becasue they would need
acces keys.
By moving them over to openfreemap (and adding a semi-minimal example of
a map which won't load) this makes the test data more realistic.

And to answer the question:
What does maplibre-native-rs do in this case?
Well, it throws an exception.