feat(pmtiles): Embed version in tilejson tiles URL (#2198)
This is more of a discussion than a PR, but I wanted to get some
feedback on the possibility of embedding version information in the
tilejson `tiles` URL. This would help a lot with cache-busting and
increasing cache hits by extending the cache age maximum.

In the case of pmtiles, I've used the technique of setting the version
in the URL to the value of the etag on the pmtiles archive, so minor
tile updates clear the cache as well -- how does that sound?

If the reception is good on this, I'd like to continue (config options?
I'm using `v` now, but if that overlaps with something else, it could
produce surprising results), but this is basically just a bare-bones
prototype to test the waters. Let me know what you think.

---------

Co-authored-by: Frank Elsinga <frank@elsinga.de>
Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>