chore(deps): sync pre-commit and biomejs versions (#2429)
We currently apparently allow more than one version of biomejs to
co-exist.
This PR fixes this and updates all versions with the one from
pre-commit.

Updating it via dependabot is already disabled.

Since only some of the version markers are updated, we need to sync them
in a similar way how we sync martins verision.

---------

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>